requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==2.4.0
Pillow==10.1.0